requirements.txt
pyarrow==17.0
joblib==1.4.2 
scikit-learn==1.5.1 
xgboost==2.1
torchdiffeq==0.2.4
optuna==3.6.1
wandb==0.17.5
pandas==2.2.2
numpy==2.0.1
matplotlib==3.9.1
seaborn==0.13.2
scipy==1.14.0
PyYAML==6.0.1
pydantic==2.8.2
torch_geometric==2.5.3
torchmetrics==1.4.0
six==1.16.0
trimap==1.1.4
umap-learn==0.5.7
pacmap==0.7.6
openTSNE==1.0.2
tensorboard==2.19.0
psycopg2-binary==2.9.10
ipython==9.0.2
imageio==2.37.0